Move matrix multiply utils to core
