+++ com.unity.render-pipelines.core/Runtime/Utilities/CoreMatrixUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d96647d3fcceac0429f12837ce7e2e3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
